requirements.txt
Jinja2==3.1.4
Markdown==3.6
MarkupSafe==3.0.3
python-slugify==8.0.4
text-unidecode==1.3
